mathQuillEditHandler outsourced to editHandler.js
